// whereTowatch/src/components/Interfaces.tsx
export interface IMovie {
  title: string;
  genre: string[];
  starring: string[];
  available: string[];
  director: string[];
  year: number;
  length: number;
  url: string;
  desc: string;
}

export interface IPoster {
  Title: string;
  Year: string;
  Rated: string;
  Released: string;
  Runtime: string;
  Genre: string;
  Director: string;
  Writer: string;
  Actors: string;
  Plot: string;
  Language: string;
  Country: string;
  Awards: string;
  Poster: string;
  Ratings: string[];
  Metascore: string;
  imdbRating: string;
  imdbVotes: string;
  imdbID: string;
  Type: string;
  DVD: string;
  BoxOffice: string;
  Production: string;
  Website: string;
  Response: string;
}